requirements.txt
absl-py==0.1.10
appnope==0.1.0
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bleach==1.5.0
bokeh==0.12.14
boto==2.48.0
boto3==1.5.30
botocore==1.8.44
bz2file==0.98
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.27.3
cytoolz==0.8.2
decorator==4.2.1
dill==0.2.7.1
dnspython==1.15.0
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
eventlet==0.24.1
fbprophet==0.2.1
Flask==1.0.2
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
ftfy==4.4.3
future==0.17.1
gensim==3.3.0
gevent==1.3.7
gevent-websocket==0.10.1
greenlet==0.4.15
html5lib==0.9999999
idna==2.7
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.4
lxml==3.4.2
malmo==0.35.6.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
monotonic==1.5
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.4.0
numpy==1.14.0
oauthlib==2.1.0
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
pexpect==4.4.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
pyparsing==2.2.0
pyrtlsdr==0.2.9
PySocks==1.6.8
pystan==2.17.1.0
python-dateutil==2.6.1
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
requests-oauthlib==1.0.0
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.6
sounddevice==0.3.12
spacy==2.0.7
SQLAlchemy==1.2.12
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.1
thinc==6.10.2
toolz==0.9.0
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
tweepy==3.6.0
ujson==1.35
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
wikipedia==1.4.0
wrapt==1.10.11
xlrd==1.1.0
